working copy of moving feed activity functions to main activity, needs to fix the states and the page layout
